chore: Rename directories
* app 디렉토리를 domain으로 변경
* infrastrucutre 디렉토리를 global으로 변경
