+++ src/main/java/leets/crazyform/domain/user/type/Vendor.java
@@ -1,4 +1,4 @@
-package leets.crazyform.app.user.type;
+package leets.crazyform.domain.user.type;
 
 public enum Vendor {
     KAKAO, GOOGLE, NAVER, FACEBOOK
