+++ src/main/java/leets/crazyform/global/error/ErrorCode.java
@@ -1,0 +1,15 @@
+package leets.crazyform.global.error;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public enum ErrorCode {
+    INTERNAL_SERVER_ERROR(500, "INTERNAL_SERVER_ERROR", "서버 오류가 발생했습니다."),
+    USER_NOT_FOUND(404, "USER_NOT_FOUND", "존재하지 않는 사용자입니다.");
+
+    private final int httpStatus;
+    private final String code;
+    private final String message;
+}
