--- src/main/java/leets/crazyform/infrastructure/error/ErrorCode.java
@@ -1,14 +1,0 @@
-package leets.crazyform.infrastructure.error;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public enum ErrorCode {
-    INTERNAL_SERVER_ERROR(500, "INTERNAL_SERVER_ERROR", "서버 오류가 발생했습니다.");
-
-    private final int httpStatus;
-    private final String code;
-    private final String message;
-}
